VideoScript/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
faster-whisper==1.0.1
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.23.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
inaSpeechSegmenter==0.7.6
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.3.1
lightning-utilities==0.11.3.post0
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
munkres==1.1.4
namex==0.0.8
networkx==3.3
nltk==3.8.1
noisereduce==3.0.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.1
opt-einsum==3.3.0
optree==0.12.0
optuna==3.6.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
protobuf==4.25.3
pyannote.algorithms==0.8
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.parser==0.8
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyparsing==3.1.2
pyppeteer==2.0.0
Pyro4==4.82
PySocks==1.7.1
pytextgrid==0.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.3.1
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
semver==3.0.2
sentencepiece==0.2.0
serpent==1.41
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.31
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tokenizers==0.15.2
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tqdm==4.66.4
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
whisperx==3.1.5
wrapt==1.16.0
wsproto==1.2.0
xarray==2024.6.0
yarl==1.9.4
zipp==3.19.2
